+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+* [PR-99](https://github.com/ITK-Leantime/project-overview/pull/99)
+  * Show "Showing X of Y" above each view's table and next to the load-more button.
 * [PR-98](https://github.com/ITK-Leantime/project-overview/pull/98)
   * Combine Priority and Status filters with AND instead of OR.